Rename and remove javadocs from JmxRmiSerialFilter Rename classes and tests
