+++ geode-core/src/main/java/org/apache/geode/management/internal/ManagementAgent.java
@@ -134,7 +134,7 @@ public class ManagementAgent {
 
 
   public synchronized void startAgent() {
-    serialFilter.configureSerialFilterIfEmpty();
+    serialFilter.configureSerialFilter();
     loadWebApplications();
 
     if (!this.running && this.config.getJmxManagerPort() != 0) {
