+++ geode-core/src/test/java/org/apache/geode/management/internal/JmxRmiOpenTypesSerialFilterTest.java
@@ -41,7 +41,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.contrib.java.lang.system.RestoreSystemProperties;
 
-public class OpenJmxTypesSerialFilterTest {
+public class JmxRmiOpenTypesSerialFilterTest {
 
   private Consumer<String> infoLogger;
 
@@ -55,7 +55,7 @@ public class OpenJmxTypesSerialFilterTest {
 
   @Test
   public void propertyValueIsNullByDefault() {
-    String propertyValue = System.getProperty(OpenJmxTypesSerialFilter.PROPERTY_NAME);
+    String propertyValue = System.getProperty(JmxRmiOpenTypesSerialFilter.PROPERTY_NAME);
 
     assertThat(propertyValue).isNull();
   }
@@ -64,9 +64,9 @@ public class OpenJmxTypesSerialFilterTest {
   public void supportsDedicatedSerialFilter_returnsTrue_atLeastJava9() {
     assumeThat(isJavaVersionAtLeast(JavaVersion.JAVA_9)).isTrue();
 
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    boolean result = openJmxTypesSerialFilter.supportsDedicatedSerialFilter();
+    boolean result = jmxRmiOpenTypesSerialFilter.supportsDedicatedSerialFilter();
 
     assertThat(result).isTrue();
   }
@@ -75,9 +75,9 @@ public class OpenJmxTypesSerialFilterTest {
   public void supportsDedicatedSerialFilter_returnsFalse_atMostJava8() {
     assumeThat(isJavaVersionAtMost(JavaVersion.JAVA_1_8)).isTrue();
 
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    boolean result = openJmxTypesSerialFilter.supportsDedicatedSerialFilter();
+    boolean result = jmxRmiOpenTypesSerialFilter.supportsDedicatedSerialFilter();
 
     assertThat(result).isFalse();
   }
@@ -85,36 +85,36 @@ public class OpenJmxTypesSerialFilterTest {
   @Test
   public void setPropertyValue_setsValue_ifExistingValueIsNull() {
     String value = "value-of-property";
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    openJmxTypesSerialFilter.setPropertyValueUnlessExists(value);
+    jmxRmiOpenTypesSerialFilter.setPropertyValueUnlessExists(value);
 
-    String propertyValue = System.getProperty(OpenJmxTypesSerialFilter.PROPERTY_NAME);
+    String propertyValue = System.getProperty(JmxRmiOpenTypesSerialFilter.PROPERTY_NAME);
     assertThat(propertyValue).isEqualTo(value);
   }
 
   @Test
   public void setPropertyValue_setsValue_ifExistingValueIsEmpty() {
-    System.setProperty(OpenJmxTypesSerialFilter.PROPERTY_NAME, "");
+    System.setProperty(JmxRmiOpenTypesSerialFilter.PROPERTY_NAME, "");
     String value = "value-of-property";
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    openJmxTypesSerialFilter.setPropertyValueUnlessExists(value);
+    jmxRmiOpenTypesSerialFilter.setPropertyValueUnlessExists(value);
 
-    String propertyValue = System.getProperty(OpenJmxTypesSerialFilter.PROPERTY_NAME);
+    String propertyValue = System.getProperty(JmxRmiOpenTypesSerialFilter.PROPERTY_NAME);
     assertThat(propertyValue).isEqualTo(value);
   }
 
   @Test
   public void setPropertyValue_logsMessage_ifExistingValueIsEmpty() {
-    System.setProperty(OpenJmxTypesSerialFilter.PROPERTY_NAME, "");
+    System.setProperty(JmxRmiOpenTypesSerialFilter.PROPERTY_NAME, "");
     String value = "value-of-property";
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter =
-        new OpenJmxTypesSerialFilter(infoLogger, () -> true);
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter =
+        new JmxRmiOpenTypesSerialFilter(infoLogger, () -> true);
 
-    openJmxTypesSerialFilter.setPropertyValueUnlessExists(value);
+    jmxRmiOpenTypesSerialFilter.setPropertyValueUnlessExists(value);
 
-    String expectedLogMessage = "System property " + OpenJmxTypesSerialFilter.PROPERTY_NAME +
+    String expectedLogMessage = "System property " + JmxRmiOpenTypesSerialFilter.PROPERTY_NAME +
         " is now configured with '" + value + "'.";
     verify(infoLogger).accept(expectedLogMessage);
   }
@@ -122,230 +122,230 @@ public class OpenJmxTypesSerialFilterTest {
   @Test
   public void setPropertyValue_leavesExistingValue_ifExistingValueIsNotEmpty() {
     String existingValue = "existing-value-of-property";
-    System.setProperty(OpenJmxTypesSerialFilter.PROPERTY_NAME, existingValue);
+    System.setProperty(JmxRmiOpenTypesSerialFilter.PROPERTY_NAME, existingValue);
     String value = "value-of-property";
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    openJmxTypesSerialFilter.setPropertyValueUnlessExists(value);
+    jmxRmiOpenTypesSerialFilter.setPropertyValueUnlessExists(value);
 
-    String propertyValue = System.getProperty(OpenJmxTypesSerialFilter.PROPERTY_NAME);
+    String propertyValue = System.getProperty(JmxRmiOpenTypesSerialFilter.PROPERTY_NAME);
     assertThat(propertyValue).isEqualTo(existingValue);
   }
 
   @Test
   public void setPropertyValue_logsMessage_ifExistingValueIsNotEmpty() {
     String existingValue = "existing-value-of-property";
-    System.setProperty(OpenJmxTypesSerialFilter.PROPERTY_NAME, existingValue);
+    System.setProperty(JmxRmiOpenTypesSerialFilter.PROPERTY_NAME, existingValue);
     String value = "value-of-property";
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter =
-        new OpenJmxTypesSerialFilter(infoLogger, () -> true);
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter =
+        new JmxRmiOpenTypesSerialFilter(infoLogger, () -> true);
 
-    openJmxTypesSerialFilter.setPropertyValueUnlessExists(value);
+    jmxRmiOpenTypesSerialFilter.setPropertyValueUnlessExists(value);
 
-    String expectedLogMessage = "System property " + OpenJmxTypesSerialFilter.PROPERTY_NAME +
+    String expectedLogMessage = "System property " + JmxRmiOpenTypesSerialFilter.PROPERTY_NAME +
         " is already configured.";
     verify(infoLogger).accept(expectedLogMessage);
   }
 
   @Test
   public void createSerialFilterPattern_includesBoolean() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    String result = openJmxTypesSerialFilter.createSerialFilterPattern();
+    String result = jmxRmiOpenTypesSerialFilter.createSerialFilterPattern();
 
     assertThat(result).contains(Boolean.class.getName());
   }
 
   @Test
   public void createSerialFilterPattern_includesByte() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    String result = openJmxTypesSerialFilter.createSerialFilterPattern();
+    String result = jmxRmiOpenTypesSerialFilter.createSerialFilterPattern();
 
     assertThat(result).contains(Byte.class.getName());
   }
 
   @Test
   public void createSerialFilterPattern_includesCharacter() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    String result = openJmxTypesSerialFilter.createSerialFilterPattern();
+    String result = jmxRmiOpenTypesSerialFilter.createSerialFilterPattern();
 
     assertThat(result).contains(Character.class.getName());
   }
 
   @Test
   public void createSerialFilterPattern_includesShort() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    String result = openJmxTypesSerialFilter.createSerialFilterPattern();
+    String result = jmxRmiOpenTypesSerialFilter.createSerialFilterPattern();
 
     assertThat(result).contains(Short.class.getName());
   }
 
   @Test
   public void createSerialFilterPattern_includesInteger() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    String result = openJmxTypesSerialFilter.createSerialFilterPattern();
+    String result = jmxRmiOpenTypesSerialFilter.createSerialFilterPattern();
 
     assertThat(result).contains(Integer.class.getName());
   }
 
   @Test
   public void createSerialFilterPattern_includesLong() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    String result = openJmxTypesSerialFilter.createSerialFilterPattern();
+    String result = jmxRmiOpenTypesSerialFilter.createSerialFilterPattern();
 
     assertThat(result).contains(Long.class.getName());
   }
 
   @Test
   public void createSerialFilterPattern_includesFloat() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    String result = openJmxTypesSerialFilter.createSerialFilterPattern();
+    String result = jmxRmiOpenTypesSerialFilter.createSerialFilterPattern();
 
     assertThat(result).contains(Float.class.getName());
   }
 
   @Test
   public void createSerialFilterPattern_includesDouble() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    String result = openJmxTypesSerialFilter.createSerialFilterPattern();
+    String result = jmxRmiOpenTypesSerialFilter.createSerialFilterPattern();
 
     assertThat(result).contains(Double.class.getName());
   }
 
   @Test
   public void createSerialFilterPattern_includesString() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    String result = openJmxTypesSerialFilter.createSerialFilterPattern();
+    String result = jmxRmiOpenTypesSerialFilter.createSerialFilterPattern();
 
     assertThat(result).contains(String.class.getName());
   }
 
   @Test
   public void createSerialFilterPattern_includesBigInteger() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    String result = openJmxTypesSerialFilter.createSerialFilterPattern();
+    String result = jmxRmiOpenTypesSerialFilter.createSerialFilterPattern();
 
     assertThat(result).contains(BigInteger.class.getName());
   }
 
   @Test
   public void createSerialFilterPattern_includesBigDecimal() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    String result = openJmxTypesSerialFilter.createSerialFilterPattern();
+    String result = jmxRmiOpenTypesSerialFilter.createSerialFilterPattern();
 
     assertThat(result).contains(BigDecimal.class.getName());
   }
 
   @Test
   public void createSerialFilterPattern_includesObjectName() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    String result = openJmxTypesSerialFilter.createSerialFilterPattern();
+    String result = jmxRmiOpenTypesSerialFilter.createSerialFilterPattern();
 
     assertThat(result).contains(ObjectName.class.getName());
   }
 
   @Test
   public void createSerialFilterPattern_includesCompositeData() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    String result = openJmxTypesSerialFilter.createSerialFilterPattern();
+    String result = jmxRmiOpenTypesSerialFilter.createSerialFilterPattern();
 
     assertThat(result).contains(CompositeData.class.getName());
   }
 
   @Test
   public void createSerialFilterPattern_includesTabularData() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    String result = openJmxTypesSerialFilter.createSerialFilterPattern();
+    String result = jmxRmiOpenTypesSerialFilter.createSerialFilterPattern();
 
     assertThat(result).contains(TabularData.class.getName());
   }
 
   @Test
   public void createSerialFilterPattern_includesSimpleType() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    String result = openJmxTypesSerialFilter.createSerialFilterPattern();
+    String result = jmxRmiOpenTypesSerialFilter.createSerialFilterPattern();
 
     assertThat(result).contains(SimpleType.class.getName());
   }
 
   @Test
   public void createSerialFilterPattern_includesCompositeType() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    String result = openJmxTypesSerialFilter.createSerialFilterPattern();
+    String result = jmxRmiOpenTypesSerialFilter.createSerialFilterPattern();
 
     assertThat(result).contains(CompositeType.class.getName());
   }
 
   @Test
   public void createSerialFilterPattern_includesTabularType() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    String result = openJmxTypesSerialFilter.createSerialFilterPattern();
+    String result = jmxRmiOpenTypesSerialFilter.createSerialFilterPattern();
 
     assertThat(result).contains(TabularType.class.getName());
   }
 
   @Test
   public void createSerialFilterPattern_includesArrayType() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    String result = openJmxTypesSerialFilter.createSerialFilterPattern();
+    String result = jmxRmiOpenTypesSerialFilter.createSerialFilterPattern();
 
     assertThat(result).contains(ArrayType.class.getName());
   }
 
   @Test
   public void createSerialFilterPattern_includesMarshalledObject() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    String result = openJmxTypesSerialFilter.createSerialFilterPattern();
+    String result = jmxRmiOpenTypesSerialFilter.createSerialFilterPattern();
 
     assertThat(result).contains(MarshalledObject.class.getName());
   }
 
   @Test
   public void createSerialFilterPattern_rejectsAllOtherTypes() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter = new OpenJmxTypesSerialFilter();
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter = new JmxRmiOpenTypesSerialFilter();
 
-    String result = openJmxTypesSerialFilter.createSerialFilterPattern();
+    String result = jmxRmiOpenTypesSerialFilter.createSerialFilterPattern();
 
     assertThat(result).endsWith("!*");
   }
 
   @Test
   public void configureSerialFilterIfEmpty_setsPropertyValue_atLeastJava9() {
-    OpenJmxTypesSerialFilter openJmxTypesSerialFilter =
-        new OpenJmxTypesSerialFilter(infoLogger, () -> true);
+    JmxRmiOpenTypesSerialFilter jmxRmiOpenTypesSerialFilter =
+        new JmxRmiOpenTypesSerialFilter(infoLogger, () -> true);
 
-    openJmxTypesSerialFilter.configureSerialFilterIfEmpty();
+    jmxRmiOpenTypesSerialFilter.configureSerialFilter();
 
-    String propertyValue = System.getProperty(OpenJmxTypesSerialFilter.PROPERTY_NAME);
+    String propertyValue = System.getProperty(JmxRmiOpenTypesSerialFilter.PROPERTY_NAME);
     assertThat(propertyValue).isNotEmpty();
   }
 
   @Test
   public void configureSerialFilterIfEmpty_setsPropertyValue_atMostJava8() {
     JmxRmiSerialFilter jmxRmiOpenTypesSerialFilter =
-        new OpenJmxTypesSerialFilter(infoLogger, () -> false);
+        new JmxRmiOpenTypesSerialFilter(infoLogger, () -> false);
 
-    jmxRmiOpenTypesSerialFilter.configureSerialFilterIfEmpty();
+    jmxRmiOpenTypesSerialFilter.configureSerialFilter();
 
-    String propertyValue = System.getProperty(OpenJmxTypesSerialFilter.PROPERTY_NAME);
+    String propertyValue = System.getProperty(JmxRmiOpenTypesSerialFilter.PROPERTY_NAME);
     assertThat(propertyValue).isNull();
   }
 }
